refactor - still infinite looping
